rename files to match the class name
