+++ lib/error_checker.rb
